Refactor InstructionTranslator and archive old stub
Reworked InstructionTranslator to use async instruction handlers and improved ARM/MIPS instruction handling logic. Archived the old stub version as InstructionTranslator_Archive.cs for reference. Minor attribute cleanup in CortexA15Cpu.cs. Updated project file to exclude the archived stub from compilation.
